StreamingTemplateRenderer: don't require layout to stream
It is entirely possible to call `render layout: false, stream: true`
for a TurboStreamer template. If `layout` is `false` then `render_template`
never returns a response which Unicorn can stream in chunks to the client.

As long as the template's handler is `TurboStreamer::Handler` then
we can build a streaming response.